+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.48.0-beta.0] - Unreleased
+## [0.49.0-beta.0] - Unreleased
 
 ### Added
 - The log file existence is checked and necessary folder structure is created if needed.
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - The support for most old `typing.<Type>` types for `configstruct` wrapper.
 
+## [0.48.0] - Accidental tag push for release, release removed
 
 ## [0.47.0] - 2025-03-14